debug message: print uint8 channel as integer not char

+++ libs/ardour/midi_port.cc
@@ -158,13 +158,13 @@ MidiPort::resolve_notes (void* port_buffer, MidiBuffer::TimeType when)
 		 */
 
 		if (port_engine.midi_event_put (port_buffer, when, ev, 3) != 0) {
-			cerr << "failed to deliver sustain-zero on channel " << channel << " on port " << name() << endl;
+			cerr << "failed to deliver sustain-zero on channel " << (int)channel << " on port " << name() << endl;
 		}
 
 		ev[1] = MIDI_CTL_ALL_NOTES_OFF;
 
 		if (port_engine.midi_event_put (port_buffer, 0, ev, 3) != 0) {
-			cerr << "failed to deliver ALL NOTES OFF on channel " << channel << " on port " << name() << endl;
+			cerr << "failed to deliver ALL NOTES OFF on channel " << (int)channel << " on port " << name() << endl;
 		}
 	}
 }
